[test/gurka/test_precision.cc] sign-compare: int -> size_t

+++ test/gurka/test_precision.cc
@@ -35,7 +35,7 @@ TEST_F(Precision, WaypointsOnNodes) {
                                         map.nodes["D"], map.nodes["E"], map.nodes["F"]};
 
   EXPECT_EQ(shape.size(), expected_shape.size());
-  for (int i = 0; i < shape.size(); ++i) {
+  for (size_t i = 0; i < shape.size(); ++i) {
     EXPECT_NEAR(shape[i].lat(), expected_shape[i].lat(), 0.000001);
     EXPECT_NEAR(shape[i].lng(), expected_shape[i].lng(), 0.000001);
   }
@@ -49,7 +49,7 @@ TEST_F(Precision, PartialOffsetCheckOne) {
                                         map.nodes["D"], map.nodes["E"], map.nodes["2"]};
 
   EXPECT_EQ(shape.size(), expected_shape.size());
-  for (int i = 0; i < shape.size(); ++i) {
+  for (size_t i = 0; i < shape.size(); ++i) {
     EXPECT_NEAR(shape[i].lat(), expected_shape[i].lat(), 0.000001);
     EXPECT_NEAR(shape[i].lng(), expected_shape[i].lng(), 0.000001);
   }
@@ -63,7 +63,7 @@ TEST_F(Precision, PartialOffsetCheckTwo) {
                                         map.nodes["D"], map.nodes["E"], map.nodes["2"]};
 
   EXPECT_EQ(shape.size(), expected_shape.size());
-  for (int i = 0; i < shape.size(); ++i) {
+  for (size_t i = 0; i < shape.size(); ++i) {
     EXPECT_NEAR(shape[i].lat(), expected_shape[i].lat(), 0.000001);
     EXPECT_NEAR(shape[i].lng(), expected_shape[i].lng(), 0.000001);
   }
@@ -77,7 +77,7 @@ TEST_F(Precision, PartialOffsetCheckThree) {
                                         map.nodes["D"], map.nodes["E"], map.nodes["F"]};
 
   EXPECT_EQ(shape.size(), expected_shape.size());
-  for (int i = 0; i < shape.size(); ++i) {
+  for (size_t i = 0; i < shape.size(); ++i) {
     EXPECT_NEAR(shape[i].lat(), expected_shape[i].lat(), 0.000001);
     EXPECT_NEAR(shape[i].lng(), expected_shape[i].lng(), 0.000001);
   }
